+++ packages/webpack-plugin-extract-css-assets/src/cjs.js
